Centralize Rails extensions; clean up utils module
