+++ lib/acts-as-taggable-on.rb
@@ -35,8 +35,6 @@ end
 
 require "acts_as_taggable_on/utils"
 
-require "acts_as_taggable_on/taggable"
-require "acts_as_taggable_on/acts_as_taggable_on/compatibility"
 require "acts_as_taggable_on/acts_as_taggable_on/core"
 require "acts_as_taggable_on/acts_as_taggable_on/collection"
 require "acts_as_taggable_on/acts_as_taggable_on/cache"
@@ -44,18 +42,7 @@ require "acts_as_taggable_on/acts_as_taggable_on/ownership"
 require "acts_as_taggable_on/acts_as_taggable_on/related"
 require "acts_as_taggable_on/acts_as_taggable_on/dirty"
 
-require "acts_as_taggable_on/tagger"
 require "acts_as_taggable_on/tag"
 require "acts_as_taggable_on/tag_list"
-require "acts_as_taggable_on/tags_helper"
 require "acts_as_taggable_on/tagging"
-require 'acts_as_taggable_on/engine'
-
-ActiveSupport.on_load(:active_record) do
-  extend ActsAsTaggableOn::Compatibility
-  extend ActsAsTaggableOn::Taggable
-  include ActsAsTaggableOn::Tagger
-end
-ActiveSupport.on_load(:action_view) do
-  include ActsAsTaggableOn::TagsHelper
-end
+require 'acts_as_taggable_on/rails_ext'
