+++ lib/acts_as_taggable_on/rails_ext.rb
@@ -1,0 +1,13 @@
+require 'acts_as_taggable_on/engine'
+require "acts_as_taggable_on/taggable"
+require "acts_as_taggable_on/acts_as_taggable_on/compatibility"
+require "acts_as_taggable_on/tagger"
+require "acts_as_taggable_on/tags_helper"
+ActiveSupport.on_load(:active_record) do
+  extend ActsAsTaggableOn::Compatibility
+  extend ActsAsTaggableOn::Taggable
+  include ActsAsTaggableOn::Tagger
+end
+ActiveSupport.on_load(:action_view) do
+  include ActsAsTaggableOn::TagsHelper
+end
